Add model and controller files
